fix: seadb vector float64 -> float32

+++ seahub/seadb_models/models.py
@@ -49,7 +49,7 @@ class PropertyTypes:
 
 class ListTypes:
     vector = {
-        "list_type": "float64",
+        "list_type": "float32",
     }
 
 class SelectTypes:
